Moved maze generation into sub mods
