+++ src/mazebuilder/mod.rs
@@ -1,0 +1,4 @@
+extern crate rand;
+mod maze;
+pub mod dfs;
+pub mod kruskal;
